server/pyproject.toml
[project]
name = "msgraph-mcp"
version = "0.1.0"
description = "MCP Server for Microsoft Graph"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "redis>=5.0.0",
  "structlog>=24.1.0",
  "pyjwt>=2.8.0",
  "opentelemetry-api>=1.23.0",
  "opentelemetry-sdk>=1.23.0",
  "opentelemetry-exporter-otlp-proto-http>=1.23.0",
  "opentelemetry-instrumentation-fastapi>=0.44b0",
  "opentelemetry-instrumentation-httpx>=0.44b0",
  "mcp>=0.1.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.2.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0"
]

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"
